fix: #51 do not correct file casing issues on case-insensitive filesystems

+++ Linux/install-ut2004.sh
@@ -1545,6 +1545,9 @@ installer::entrypoint() {
 
   local DESTINATION_HOMIFIED=""
 
+  # shellcheck disable=SC2034 # Used in some installers
+  local IS_DESTINATION_CASE_SENSITIVE_FS=""
+
   step::check_destination() {
     term::step::new "Checking Destination Folder"
 
@@ -1578,6 +1581,30 @@ installer::entrypoint() {
       term::step::failed_with_error "The ./Installer subfolder of the destination cannot be written by this user."
       return 77 #E_PERM
     fi
+
+    # Check Case-Sensitivity
+    {
+      if [[ -f "${_arg_destination%/}/Installer/.check_casesensitive" ]]; then
+        rm -f "${_arg_destination%/}/Installer/.check_casesensitive"
+      fi
+
+      if [[ ! -f "${_arg_destination%/}/Installer/.Check_caseSensitive" ]]; then
+        touch "${_arg_destination%/}/Installer/.Check_caseSensitive"
+      fi
+
+      if [[ -f "${_arg_destination%/}/Installer/.check_casesensitive" ]]; then
+        # shellcheck disable=SC2034 # Used in some installers
+        IS_DESTINATION_CASE_SENSITIVE_FS="no"
+      else
+        # shellcheck disable=SC2034 # Used in some installers
+        IS_DESTINATION_CASE_SENSITIVE_FS="yes"
+      fi
+
+      rm -f "${_arg_destination%/}/Installer/.Check_caseSensitive"
+    } || {
+      term::step::failed_with_error "Unable to determine case sensitivity. Aborting installation."
+      return 77 #E_PERM
+    }
 
     DESTINATION_HOMIFIED="${_arg_destination}"
     { [[ "${DESTINATION_HOMIFIED}" =~ ^"${HOME}"(/|$) ]] && DESTINATION_HOMIFIED="~${_arg_destination#"${HOME}"}"; } || true
@@ -2103,29 +2130,32 @@ You may read the Terms of Service at this URL:
 
     local SYSTEM_FOLDER="${_arg_destination%/}/System${UE_SYSTEM_FOLDER_SUFFIX}"
 
-    # Remove files with different casing than the patch payload
-    COMMON_WRONG_CASINGS=(
-      "Bonuspack.u"
-      "Gui2K4.u"
-      "Gameplay.u"
-      "Ipdrv.u"
-      "Skaarjpack.u"
-      "StreamLineFX.u"
-      "UT2K4Assault.u"
-      "UT2K4AssaultFull.u"
-      "XVoting.u"
-      "xWebAdmin.u"
-    )
+    if [[ "${IS_DESTINATION_CASE_SENSITIVE_FS}" == "yes" ]]; then
+      # Remove files with different casing than the patch payload
+      local WRONG_CASING
+      local COMMON_WRONG_CASINGS=(
+        "Bonuspack.u"
+        "Gui2K4.u"
+        "Gameplay.u"
+        "Ipdrv.u"
+        "Skaarjpack.u"
+        "StreamLineFX.u"
+        "UT2K4Assault.u"
+        "UT2K4AssaultFull.u"
+        "XVoting.u"
+        "xWebAdmin.u"
+      )
 
-    for WRONG_CASING in "${COMMON_WRONG_CASINGS[@]}"; do
-      if [[ -f "${SYSTEM_FOLDER}/${WRONG_CASING}" ]]; then
-        rm -f "${SYSTEM_FOLDER}/${WRONG_CASING}"
-      fi
+      for WRONG_CASING in "${COMMON_WRONG_CASINGS[@]}"; do
+        if [[ -f "${SYSTEM_FOLDER}/${WRONG_CASING}" ]]; then
+          rm -f "${SYSTEM_FOLDER}/${WRONG_CASING}"
+        fi
 
-      if [[ -f "${_arg_destination%/}/System/${WRONG_CASING}" ]]; then
-        rm -f "${_arg_destination%/}/System/${WRONG_CASING}"
-      fi
-    done
+        if [[ -f "${_arg_destination%/}/System/${WRONG_CASING}" ]]; then
+          rm -f "${_arg_destination%/}/System/${WRONG_CASING}"
+        fi
+      done
+    fi
 
     # Remove provided libopenal if provided by the system
     if [[ -f "${SYSTEM_FOLDER}/libopenal.so.1" ]] && [[ -n "$(step::ut2004_special_fixes::find_library libopenal.so.1)" ]]; then
